Plan singleton pattern
