+++ src/main/java/com/sbancuz/plannh/nei/PlanOverlayHandler.java
@@ -11,6 +11,7 @@ import com.sbancuz.plannh.api.RecipePropertyAPI;
 import com.sbancuz.plannh.data.PropertyProvider;
 import com.sbancuz.plannh.data.flowchart.Graph;
 import com.sbancuz.plannh.data.flowchart.Node;
+import com.sbancuz.plannh.data.flowchart.Plan;
 import com.sbancuz.plannh.gui.FlowchartScreen;
 
 import codechicken.nei.PositionedStack;
@@ -53,13 +54,13 @@ public class PlanOverlayHandler implements IOverlayHandler {
 
     private static void addRecipe(final GuiContainer firstGui, final IRecipeHandler handler, final int recipeIndex) {
         final Node node = new Node(handler, recipeIndex, DEFAULT_NODE_X, DEFAULT_NODE_Y);
-        final Graph graph = PlanAPI.getActiveGraph();
-        graph.addNode(node);
+        final Graph graph = Plan.getActiveGraph();
+        graph.getNodes().put(node.id, node);
         PlanAPI.save();
 
         if (firstGui instanceof final GuiContainerWrapper wrapper
             && wrapper.getScreen() instanceof final FlowchartScreen screen) {
-            screen.canvas.rebuildNodeWidgets();
+//            screen.canvas.rebuildNodeWidgets();
         }
     }
